==> scripts/2a_mp_org_list.ipynb <==
#file is to clean the list of politicians and organisations
import pandas as pd

mp_df = pd.read_csv('../data/mp_list.csv')

org_df = pd.read_csv('../data/org_list.csv')

mp_df.columns = [column.lower() for column in mp_df.columns]

initials_list = []
for name in mp_df.mp:
    name_parts = name.split(' ')
    initials = ''
    for part in name_parts:
        initials += part[0]
    if len(initials)>=2:
        initials_list.append(initials)
    else: initials_list.append('')

mp_df['initials'] = initials_list

given_name_parts = [[name for name in  names.split(' ')] for names in mp_df.mp]

first_names = []
last_names = []
for names in given_name_parts:
    if len(names) == 2:
        if names[0] !='K' and names[0] !='S.':
            first_names.append(names[0])
            last_names.append(names[1])
        else:
            first_names.append(names[1])
            last_names.append('')
    elif len(names) >= 4:
        first_names.append(names[0:1][0])
        last_name = names[2: len(names)]
        last_name = " ".join(last_name)
        last_names.append(last_name[0])
    else:
        first_name = names[1: len(names)]
        first_name = " ".join(first_name)
        first_names.append(first_name)
        last_names.append(names[0])

first_names

mp_df['given_names'] = first_names
mp_df['family_names'] = last_names

aliases_list = []
for alias in mp_df.aliases:
    try:
        indiv_alias_list = []
        for indiv_alias in alias.replace("'", '').split(','):
            specific_alias = indiv_alias.strip()
            indiv_alias_list.append(specific_alias)
        #print(indiv_alias_list)
        aliases_list.append(indiv_alias_list)
    except:
        aliases_list.append([])

org_aliase_list = []
for alias in org_df.aliases:
    try:
        indiv_alias_list = []
        for indiv_alias in alias.replace("'", '').split(','):
            specific_alias = indiv_alias.strip()
            indiv_alias_list.append(specific_alias)
        org_aliase_list.append(indiv_alias_list)
    except:
        org_aliase_list.append([])

mp_df.aliases = aliases_list
org_df.aliases = org_aliase_list

mp_df.to_pickle('../data/mplist_clean.pk1')
org_df.to_pickle('../data/orglist_clean.pk1')
