==> scripts/2b_cleaning.ipynb <==
import pandas as pd

df = pd.read_csv('../data/submissions_comments.csv')
# the full file isnt available on github due to size limitations partial file with 80,000 comments

df.head()

df = df[['comment', 'user', 'date_time', 'sub_title']]

df

df_no_nulls = df.dropna(subset = 'comment')

df_no_nulls.info()

import redditcleaner

df['clean_comments'] = df.comment.map(redditcleaner.clean)

df.comment = df.clean_comments

df.to_pickle('../data/comments_clean.pk1')
#full file isnt available on git smaller sample is provided
