==> scripts/6_classifying_relevant_comments.ipynb <==
import numpy as np
import pandas as pd
import spacy
from tqdm import tqdm
from transformers import pipeline
from transformers import AutoTokenizer
from datetime import datetime as dt
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
import torch

#setting GPU usage
spacy.prefer_gpu()

# Loading Model trained with the NEL
nlp = spacy.load('../model')

torch.cuda.empty_cache()

print(torch.cuda.memory_reserved(0))
torch.cuda.get_device_properties(0).total_memory

df = pd.read_pickle('../data/comments_clean.pk1')
# the full file isnt available on github a smaller file is provided

df.reset_index(drop = True, inplace = True)
df.dropna(axis = 0, subset = 'comment', inplace = True)
df.drop_duplicates(subset = ['comment'], keep = 'first', inplace = True)

df.info()

#creating mirror df without nulls in the title
df2 = df.dropna(subset = 'sub_title', axis = 0)

unique_titles = [title  for title in df2.sub_title.unique()]

df= df[df.comment != '[deleted]']

df.info()

df['comment_len'] = [len(comment) for comment in df.comment]


df['comment_len'] = [len(comment.split()) for comment in df.comment]


df = df[df.user != 'sneakpeek_bot']

df.sort_values(by = 'comment_len', ascending=False)

title_ents = []
title_list = []
title_dict = {}
count = 0
for doc in nlp.pipe(tqdm(unique_titles)):
    doc_ents = []
    for ent in doc.ents:
        if ent.kb_id_ != 'NIL':
            ent_details = ent.kb_id_
            doc_ents.append(ent_details)
    title_ents.append(doc_ents)
    title_list.append(doc.text)

    title_dict[f'{count}'] = {'comment': doc.text}
    relevant_sections = {}
    for ents in doc.ents:
        relevant_sections[f'{ents.kb_id_}'] = []
        for token in ents:
            if token.ent_iob == 3:
                clause = {}
                current_token = token
                subject_count = 1
                while current_token.dep_ != "ROOT" and subject_count <= 1:
                    if current_token.dep_ ==  'nsubj' or current_token.dep_== 'iobj' or current_token.dep_== 'dobj' or current_token.dep_ == 'pobj':
                        subject_count += 1
                        current_token = current_token.head
                    else:
                        current_token = current_token.head
        section = [t for t in current_token.subtree]
        relevant_sections[f'{ents.kb_id_}'].append(section)
        title_dict[f'{count}']['entities'] = relevant_sections
    count +=1


torch.cuda.empty_cache()

title_df= pd.DataFrame({'title': title_list, 'title_ents': title_ents})

title_df = title_df[title_df.title_ents.map(lambda x: len(x))>0]

titles_with_ents = [titles for titles in title_df.title]


comment_ents = []

comment = []

%%time
'''
for doc in nlp.pipe(tqdm(df.clean_comments)):
    doc_ents = []
    for ent in doc.ents:
        if ent.kb_id_ != 'NIL':
            ent_details = ent.kb_id_
            doc_ents.append(ent_details)
            comment.append(doc.text)
    comment_ents.append(doc_ents)
    comment.append(doc.text)

'''

%%time
comment_dict = {}
count = 0
for doc in nlp.pipe(tqdm(df.clean_comments)):
    doc_ents_id = []
    comment_dict[f'{count}'] = {'comment': doc.text}
    comment_dict[f'{count}']['entities'] = []
    for ent in doc.ents:
        relevant_sections = {}
        if ent.kb_id_ != 'NIL':
            ent_details = ent.kb_id_
            doc_ents_id.append(ent_details)
            comment.append(doc.text)
            relevant_sections[f'{ent.kb_id_}'] = []
            current_token = ent[0]
            subject_count = 1
            while current_token.dep_ != "ROOT" and subject_count <= 1:
                if current_token.dep_ ==  'nsubj' or current_token.dep_== 'iobj' or current_token.dep_== 'dobj' or current_token.dep_ == 'pobj':
                    subject_count += 1
                    current_token = current_token.head
                else:
                    current_token = current_token.head
                section = [t for t in current_token.subtree]
            relevant_sections[f'{ent.kb_id_}'].append(section)
    comment_dict[f'{count}']['entities'].append(relevant_sections)
    count +=1
    comment_ents.append(doc_ents)
    comment.append(doc.text)

zip_data = zip(comment_ents, comment)

comment_df = pd.DataFrame(data = zip_data, columns = ['comment_entities', 'comment'])

comment_df.to_pickle('../data/comments_entities_only.pk1')

comment_dict

comment_df = pd.DataFrame(data = zip_data, columns = ['entities', 'polarity', 'subjectivity'])

comment_df['comment'] = [comment for comment in df.comment]

comment_df.info()

df_with_scores = df.merge(right = comment_df, how = 'left', right_on = 'comment', left_on = 'comment')

title_df= pd.DataFrame({'title': title_list, 'title_ents': title_ents})

df_with_scores.head()

df_with_scores.head()

df_with_scores.drop(columns = ['Unnamed: 0'] , inplace = True)

df_with_scores.info()

df_with_scores['date_time'] = [date for date in df_with_scores.date_time]

df_with_scores = df_with_scores.merge(right = title_df, how = 'left', left_on = 'sub_title', right_on = 'title')

df_with_scores['entities_both'] = np.where(df_with_scores.entities.map(lambda x: len(x)) ==0, df_with_scores.title_ents, df_with_scores.entities)

df_clean = df_with_scores.dropna(axis = 0, how = 'any', inplace = False)
df_clean.info()

df_clean.info()

df_clean['len_entities_both'] = [len(entity) for entity in df_clean.entities_both]

df_clean2 = df_clean[df_clean.len_entities_both >= 1]

df_clean2.info()

comments = [comment for comment in df_clean2.comment]

tokenizer = AutoTokenizer.from_pretrained('cardiffnlp/twitter-roberta-base-sentiment')
sentiment = pipeline(task = 'sentiment-analysis', model = 'cardiffnlp/twitter-roberta-base-sentiment', tokenizer = tokenizer, device = 0)

%%time
sents = sentiment(comments, max_length = 512, truncation = True, padding = 'max_length', batch_size = 64)

sents

df_clean2['label_h'] = [sents['label'] for sents in sents]

df_clean2['label_h'] = df_clean2['label_h'].map({'LABEL_0': -1, 'LABEL_1': 0, 'LABEL_2': 1})

df_clean2['score_h'] = [sents['score'] for sents in sents]

df_clean2['sent_score_h'] = df_clean2.score * df_clean2.label_h

df_clean2.head(30)

df_clean2.info()

df_clean2['date'] = [dt.utcfromtimestamp(time) for time in df_clean2['date_time']]
df_clean2['year_month'] = [date_time.to_period("M") for date_time in df_clean2.date]
df_clean2.year_month = [d_t.strftime('%Y-%m') for d_t in df_clean2.year_month]

df_clean2.info()

df_clean2.reset_index(inplace = True, drop = True)

df_clean2.entities = [list(set(entity)) for entity in df_clean2.entities]
df_clean2.entities_both = [list(set(entity)) for entity in df_clean2.entities_both]

df_clean2['len_entities'] = [len(entity) for entity in df_clean2.entities_both]

df_clean2

df_clean2 = df_clean2[df_clean2.len_entities >= 1]

df_clean2

entities = []
for ents in df_clean2.entities:
    for ent in ents:
        entities.append(ent)

unique_entities = sorted(set(entities))
unique_entities = [entity for entity in unique_entities]

pd.DataFrame(unique_entities, columns = ['entities']).to_csv('../data/entity_list.csv')

sorted_dates = [item for item in df_clean2.year_month.unique()]
pd.DataFrame(sorted_dates, columns = ['month']).sort_values(by= 'month', ascending=True).to_csv('../data/date_list.csv')

df_clean2['label_t'] =np.where(df_clean2.polarity < -0.05, '-1',
                            np.where(df_clean2.polarity >0.05, '1', '0'))

analyzer = SentimentIntensityAnalyzer()

df_clean2['output_v'] = [analyzer.polarity_scores(comment) for comment in tqdm(df_clean2.comment)]

df_clean2['score_v'] = [output['compound'] for output in df_clean2.output_v]

df_clean2['label_v'] = np.where(df_clean2.polarity < -0.05, '-1',
                               np.where(df_clean2.polarity >0.05, '1', '0'))

df_clean2.reset_index(inplace = True, drop = True)

for item in tqdm(unique_entities):
    rows = []
    for index,row in df_clean2.iterrows():
        if item in row.entities:
            rows.append(row.values)
    item_df = pd.DataFrame(data = rows, columns = df_clean2.columns)
    item_df.entities = item
    item_df.to_pickle(f'../data/indiv_data/{item}.pk1')

df_clean2.to_pickle('../data/label_data.pk1')
